packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aif-core"
version = "0.1.0"
description = "Adaptive Image Format — Server-side image processing pipeline for dark/light mode adaptation"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "Markus Köplin" },
]
keywords = ["image-processing", "dark-mode", "adaptive", "xmp", "avif"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "scikit-learn>=1.3",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/MKsys1337/AIF"
Repository = "https://github.com/MKsys1337/AIF"
Documentation = "https://github.com/MKsys1337/AIF/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/aif"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
